refactor(uri): rename files
